requirements.txt
pytest
requests
allure-pytest
pyyaml